+++ js/components/urlBarSuggestions.js
@@ -13,11 +13,13 @@ const top500 = require('./../data/top500.js')
 const {isSourceAboutUrl, isUrl} = require('../lib/appUrlUtil')
 const Immutable = require('immutable')
 const debounce = require('../lib/debounce.js')
-const {getSiteIconClass} = require('../state/siteUtil.js')
 const settings = require('../constants/settings')
 const siteTags = require('../constants/siteTags')
+const suggestionTypes = require('../constants/suggestionTypes')
 const getSetting = require('../settings').getSetting
 const eventUtil = require('../lib/eventUtil.js')
+const cx = require('../lib/classSet.js')
+const locale = require('../l10n')
 
 class UrlBarSuggestions extends ImmutableComponent {
   constructor (props) {
@@ -88,17 +90,64 @@ class UrlBarSuggestions extends ImmutableComponent {
     // If there is a URL suffix that means there's an active autocomplete for the first element.
     // We should show that as selected so the user knows what is being matched.
     const hasUrlSuffix = this.props.activeFrameProps.getIn(['navbar', 'urlbar', 'suggestions', 'urlSuffix']) || ''
-    return <ul className='urlBarSuggestions'>
-      {suggestions.map((suggestion, index) =>
-        <li data-index={index + 1}
+
+    const tabSuggestions = suggestions.filter((s) => s.type === suggestionTypes.TAB)
+    const bookmarkSuggestions = suggestions.filter((s) => s.type === suggestionTypes.BOOKMARK)
+    const historySuggestions = suggestions.filter((s) => s.type === suggestionTypes.HISTORY)
+    const searchSuggestions = suggestions.filter((s) => s.type === suggestionTypes.SEARCH)
+    const topSiteSuggestions = suggestions.filter((s) => s.type === suggestionTypes.TOP_SITE)
+
+    let items = []
+    let index = 0
+    const addToItems = (suggestions, sectionKey, title, icon) => {
+      if (suggestions.size > 0) {
+        items.push(<li className='suggestionSection'>
+          <span className='suggestionSectionTitle'>{title}</span>
+          {
+            icon
+            ? <span className={cx({
+              suggestionSectionIcon: true,
+              [sectionKey]: true,
+              fa: true,
+              [icon]: true
+            })} />
+            : null
+          }
+        </li>)
+      }
+      items = items.concat(suggestions.map((suggestion, i) => {
+        const currentIndex = index + i
+        return <li data-index={currentIndex + 1}
           onMouseOver={this.onMouseOver.bind(this)}
           onClick={suggestion.onClick}
           key={suggestion.title}
-          className={this.activeIndex === index + 1 || index === 0 && hasUrlSuffix ? 'selected' : ''}>
-          <span className={`suggestionIcon fa ${suggestion.iconClass}`} />
-          <span className='suggestionText'>{suggestion.title}</span>
+          className={cx({
+            selected: this.activeIndex === currentIndex + 1 || currentIndex === 0 && hasUrlSuffix,
+            suggestionItem: true,
+            [suggestion.type]: true
+          })}>
+          {
+            suggestion.type !== suggestionTypes.TOP_SITE
+            ? <div className='suggestionTitle'>{suggestion.title}</div>
+            : null
+          }
+          {
+            suggestion.type !== suggestionTypes.SEARCH
+            ? <div className='suggestionLocation'>{suggestion.location}</div>
+            : null
+          }
         </li>
-      )}
+      }))
+      index += suggestions.size
+    }
+    addToItems(tabSuggestions, 'tabsTitle', locale.translation('tabsSuggestionTitle'), 'fa-external-link')
+    addToItems(bookmarkSuggestions, 'bookmarksTitle', locale.translation('bookmarksSuggestionTitle'), 'fa-star-o')
+    addToItems(historySuggestions, 'historyTitle', locale.translation('historySuggestionTitle'), 'fa-clock-o')
+    addToItems(searchSuggestions, 'searchTitle', locale.translation('searchSuggestionTitle'), 'fa-search')
+    addToItems(topSiteSuggestions, 'topSiteTitle', locale.translation('topSiteSuggestionTitle'), 'fa-link')
+
+    return <ul className='urlBarSuggestions'>
+      {items}
     </ul>
   }
 
@@ -145,7 +194,7 @@ class UrlBarSuggestions extends ImmutableComponent {
     const urlLocationLower = this.props.urlLocation.toLowerCase()
     let suggestions = new Immutable.List()
     const defaultme = (x) => x
-    const mapListToElements = ({data, maxResults, classHandler, clickHandler = navigateClickHandler,
+    const mapListToElements = ({data, maxResults, type, clickHandler = navigateClickHandler,
         sortHandler = defaultme, formatTitle = defaultme, formatUrl = defaultme,
         filterValue = (site) => site.toLowerCase().includes(urlLocationLower)
     }) => // Filter out things which are already in our own list at a smaller index
@@ -161,7 +210,8 @@ class UrlBarSuggestions extends ImmutableComponent {
         return {
           onClick: clickHandler.bind(null, site),
           title: formatTitle(site),
-          location: formatUrl(site)
+          location: formatUrl(site),
+          type
         }
       })
 
@@ -170,7 +220,7 @@ class UrlBarSuggestions extends ImmutableComponent {
       suggestions = suggestions.concat(mapListToElements({
         data: this.props.frames,
         maxResults: config.urlBarSuggestions.maxOpenedFrames,
-        classHandler: () => 'fa-file',
+        type: suggestionTypes.TAB,
         clickHandler: (frameProps) =>
           windowActions.setActiveFrame(frameProps),
         formatTitle: (frame) => frame.get('title') || frame.get('location'),
@@ -185,8 +235,8 @@ class UrlBarSuggestions extends ImmutableComponent {
     if (getSetting(settings.BOOKMARK_SUGGESTIONS)) {
       suggestions = suggestions.concat(mapListToElements({
         data: this.props.sites,
-        maxResults: config.urlBarSuggestions.maxSites,
-        classHandler: getSiteIconClass,
+        maxResults: config.urlBarSuggestions.maxBookmarkSites,
+        type: suggestionTypes.BOOKMARK,
         clickHandler: navigateClickHandler((site) => {
           return site.get('location')
         }),
@@ -209,8 +259,8 @@ class UrlBarSuggestions extends ImmutableComponent {
     if (getSetting(settings.HISTORY_SUGGESTIONS)) {
       suggestions = suggestions.concat(mapListToElements({
         data: this.props.sites,
-        maxResults: config.urlBarSuggestions.maxSites,
-        classHandler: getSiteIconClass,
+        maxResults: config.urlBarSuggestions.maxHistorySites,
+        type: suggestionTypes.HISTORY,
         clickHandler: navigateClickHandler((site) => {
           return site.get('location')
         }),
@@ -233,8 +283,8 @@ class UrlBarSuggestions extends ImmutableComponent {
     if (getSetting(settings.OFFER_SEARCH_SUGGESTIONS)) {
       suggestions = suggestions.concat(mapListToElements({
         data: this.props.suggestions.get('searchResults'),
-        maxResults: config.urlBarSuggestions.maxTopSites,
-        classHandler: () => 'fa-search',
+        maxResults: config.urlBarSuggestions.maxSearch,
+        type: suggestionTypes.SEARCH,
         clickHandler: navigateClickHandler((searchTerms) => this.props.searchDetail.get('searchURL')
           .replace('{searchTerms}', encodeURIComponent(searchTerms)))}))
     }
@@ -242,8 +292,8 @@ class UrlBarSuggestions extends ImmutableComponent {
     // Alexa top 500
     suggestions = suggestions.concat(mapListToElements({
       data: top500,
-      maxResults: config.urlBarSuggestions.maxSearch,
-      classHandler: () => 'fa-link',
+      maxResults: config.urlBarSuggestions.maxTopSites,
+      type: suggestionTypes.TOP_SITE,
       clickHandler: navigateClickHandler((x) => x)}))
 
     return suggestions