+++ js/constants/suggestionTypes.js
@@ -1,0 +1,13 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this file,
+ * You can obtain one at http://mozilla.org/MPL/2.0/. */
+
+const suggestionTypes = {
+  TAB: 'tabSuggestion',
+  BOOKMARK: 'bookmarkSuggestion',
+  HISTORY: 'historySuggestion',
+  SEARCH: 'searchSuggestion',
+  TOP_SITE: 'topSiteSuggestion'
+}
+
+module.exports = suggestionTypes